Make File Paths Relative
The large majority of the file paths in our project are relative
(without starting '/'). This is great in the S3 bucket where all of our
files are in the root of the bucket, but this doesn't work well when
trying to test the website locally since the project isn't in the root
of my pc. This commit standardizes the file paths as relative so they
work well on both S3 and local.