Added GitIgnore to Repo
All good repos should have a project-wide gitignore file to exclude unnecessary files
from the repo such as logs, and personal settings.
Edit: just realized there was a .gitignore in the .idea folder. I made
the .gitignore in the root of the project so that it will be useful to
ignore files in other parts of the repo.