Fix Race Condition on Account Page
There was an issue in account.js where the async method
requestUserInfo() was being called and then another
async method getClickedIdInfo() was being called. This
created a race condition where occasionally requestUserInfo
would complete after getClickedIdInfo, causing the wrong
information to appear on the page. Fixing the race
condition also made the page load faster/smoother because
it is now only making a single API call and updating the
page a single time.