Make Loading Smoother
I hid the content of the page while the lambdas are loading, and then
displayed the page content all at once when it finishes loading
completely. That got rid of the weird "half loaded" pages that were
confusing for users.
Then, since the page is blank while the API calls are happening, I added
a loading spinner to let the user know something is happening.