Fix Out of Date Session Bug on Login
There was an issue where, when you try to log in, the site would
navigate to the type of profile page belonging to the user that
previously logged in. (It was one login behind). This happened
because I think I understood how to use promises incorrectly.
But now it is fixed.